requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.38
pykrx>=1.0.45
slack-sdk>=3.0.0
python-dotenv>=1.0.0
